requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
langchain>=0.0.350
scikit-learn>=1.3.0
openai>=1.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0
